some other changes

+++ backend/src/pong-game/default.service.ts
@@ -98,6 +98,7 @@ export class DefaultService {
     const playground = new PlayGround(0, 0, 800, 600, 'black', 9, false);
     first.data.playground = playground;
     second.data.playground = playground;
+    this.logger.log('Starting Game in Room: ' + roomname + ' between: ' + first.data.user.username + ' & '+ second.data.user.username);
     const timer = setInterval(() => {
       if (playground.update(/* roomname, wss */) == false) {
 
@@ -113,7 +114,7 @@ export class DefaultService {
         // game finished
         clearInterval(timer);
         clearInterval(first.data.gameInterval);
-        this.logger.log('Game in room ' + roomname + ' Finished');
+        this.logger.log('Game in Room: ' + roomname + ' between: ', first.data.user.username + ' & ' + second.data.user.username + ' Finished');
         if (playground.scoreBoard.playerOneScore > playground.scoreBoard.playerTwoScore) {
           this.usersService.updateLevel(first.data.user.id);
           this.usersService.winsGame(first.data.user.id);
@@ -134,31 +135,36 @@ export class DefaultService {
 
   async handleUserDisconnected(wss: Server, client: Socket) {
     if (client.handshake.query.role === 'player' && client.data.gameInterval) {
-      client.data.playground.ball.reset(
-        client.data.playground.width / 2,
-        client.data.playground.height / 2,
-      );
-      client.data.playground.leftPaddle.reset();
-      client.data.playground.rightPaddle.reset();
-      wss.to(client.data.roomname).emit('gameInterrupted', {
-        playground: this.handleGetBackGround(client.data.playground),
-      });
+      if (client.data.gameInterval._destroyed === false) {
+        client.data.playground.ball.reset(
+          client.data.playground.width / 2,
+          client.data.playground.height / 2,
+        );
+        client.data.playground.leftPaddle.reset();
+        client.data.playground.rightPaddle.reset();
+        wss.to(client.data.roomname).emit('gameInterrupted', {
+          playground: this.handleGetBackGround(client.data.playground),
+        });
+        // clearInterval if not destroyed
+        clearInterval(client.data.gameInterval);
+        this.logger.log('Game Interval Cleared');
 
+        // Update Level and wins and loses for both players
+        console.log(client.data.opponentId);
+        await this.usersService.updateLevel(client.data.opponentId);
+        await this.usersService.winsGame(client.data.opponentId);
+        await this.usersService.LostGame(client.data.user.id);
+        // delete room from database
+        await this.pongGameService.deleteRoom(client.data.roomname);
+        this.logger.log('Game in Room: ' + client.data.roomname + ' Finished');
+      }
+      
       // client left room
       client.leave(client.data.roomname);
-      clearInterval(client.data.gameInterval);
-
+      
       // Update Status to ONLINE again
       await this.usersService.updateStatus(client.data.user.id, UserStatus.ONLINE);
-      
-      // Update Level and wins and loses for both players
-      this.usersService.updateLevel(client.data.opponentId);
-      this.usersService.winsGame(client.data.opponentId);
-      this.usersService.LostGame(client.data.user.id);
 
-      // delete room from database
-      await this.pongGameService.deleteRoom(client.data.roomname);
-      this.logger.log('Game Interval Cleared');
     } else if (client.handshake.query.role === 'player') {
       // Update Status to ONLINE again
       await this.usersService.updateStatus(client.data.user.id, UserStatus.ONLINE);
